refactor(cketh): share size-prefixed principal-into-32-bytes encoder
Consolidate the size-prefixed principal-into-32-bytes encoding that was
duplicated across `deposit_address::principal_to_bytes32`, the
`principal_to_hex` helper binary and the test helper
`to_32_bytes_with_size_prefix` into a single `principal_to_bytes32`
paired with the existing `parse_principal_from_slice` decoder in
`eth_logs`. No behavior change.
