+++ rs/ethereum/cketh/minter/BUILD.bazel
@@ -188,6 +188,7 @@ rust_binary(
     srcs = ["bin/principal_to_hex.rs"],
     deps = [
         # Keep sorted.
+        ":minter",
         "@crate_index//:candid",
         "@crate_index//:hex",
     ],
